requirements.txt
boto3==1.35.5
botocore==1.35.5
PyYAML==6.0.2
requests==2.32.3
python_dotenv==1.0.1